hlmv push
finally got hlmv to work under x64 source sdk. Still is trying to call for script weapon txt, but this is a start. Had to revert filesystem_init, since valve fucking added a check if a tool is attempting to use appid, which is dumb. studio.h is updated to have some of the old 32bit code exposed for hlmv. I really need to use the newer code in the future tbh.