disable lua, update vpc scripts, revert studio.h, add gamepadui
lua caused signifigant crashing, no ideas as to why, vpc scripts remove it, and we don't build it anymore. Studio.h reverted, as a fix i did to build hlmv seems to have caused an issue in-game, so stdio2 needs to be used for hlmv (prob what valve did). misyl's gamepad ui has been added as well, with new code specifically to work with the L4D+ gameui in the client dll (wip).