Update content_mounter
I updated this since Half-Life 2 has changed since the 20th anniversary, and removed 2 redundent entries. Also added CS:GO, pointing towards the cs2 directory for this. L4D, L4D2, and Portal 2 was also added to the mounter. This needs to be optimiazed, as if you load too much, and attempt to skip the games intro vid, it crashs because it doesn't have time to load everything in. Vvis mentioned only loading content in that we need at a time.