[bugfix] re-add hl2mp
apparently, we are forced into loading this. if we don't, it causes a host of issues. mounting hl2 is not a solution, as theres things broken there that don't break in the sdk (cough cough gravitygun orbs). To be fair, i didn't know this was a bug at all when using this system. Also, added stanley parable to the content mounter.