particle rain + other l4d stuff reimplentation
re implanted this from the singleplayer branch of srcbox. This seems to be busted at this moment, but particle ash/rain load in from whats been tested. Havn't a clue at this moment to what is wrong. Raytrace has also been linked over to client_base for the new particle stuff. Thanks Vvis for some reference while attempting to debug the issue!